refactor(ib-partition-controller): extract into dedicated crate
Move the IB partition state controller out of `crates/api` into a new
`carbide-ib-partition-controller` crate as part of the ongoing
controller extraction work.
