+++ crates/api/src/state_controller/mod.rs
@@ -18,7 +18,6 @@
 pub mod common_services;
 pub mod dpa_interface;
 pub(crate) mod external_service_error;
-pub mod ib_partition;
 pub mod machine;
 pub mod power_shelf;
 pub mod rack;
